requirements.txt
django==1.7.4
requests
beautifulsoup4
djangorestframework
mysqlclient
django-cors-headers
django-celery
django-filter==0.11.0
geopy